Moved packages to their own repository, fixed multiple output bugs, and added package list updater

+++ src/DisplayExecute.cpp
@@ -15,9 +15,13 @@
 
 #include "../include/DisplayExecute.hpp"
 
-int DispExecute( std::string cmd, std::string & err, bool show_output )
+int DispExecute( std::string cmd, std::string & err, bool show_output, bool create_temp_file )
 {
-	std::string finalcmd = cmd + " 2>" + TMP_FILE;
+	std::string finalcmd = cmd;
+	if( create_temp_file )
+		finalcmd += " 2>" + TMP_FILE;
+	else
+		finalcmd += " 2>&1";
 
 	FILE * pipe = popen( finalcmd.c_str(), "r" );
 
@@ -26,18 +30,15 @@ int DispExecute( std::string cmd, std::string & err, bool show_output )
 	if( !pipe )
 		return false;
 
-	winsize w;
-	ioctl( STDOUT_FILENO, TIOCGWINSZ, & w );
-
-	// To accomodate \n, max column usable is ws_col - 1
-	int term_width = w.ws_col - 1;
-
 	int current_disp_len = GetLastDispLen();
 
 	std::array< char, 10000 > opline_temp;
 
 	std::string brackets = "[  ]";
 	std::string continuation = " ...";
+
+	winsize w;
+	int term_width;
 
 	while( !feof( pipe ) ) {
 		if( fgets( opline_temp.data(), 10000, pipe ) != NULL && show_output ) {
@@ -46,15 +47,25 @@ int DispExecute( std::string cmd, std::string & err, bool show_output )
 
 			TrimString( opline );
 
+			ioctl( STDOUT_FILENO, TIOCGWINSZ, & w );
+			// To accomodate \n, max column usable is ws_col - 1
+			term_width = w.ws_col - 1;
+
 			if( current_disp_len + brackets.size() + opline.size() >= term_width ) {
 				// the size to work with is:
-				// termwidth - 1 -> total size to work with
+				// term_width -> total size to work with
 				// - current_disp_len -> already used by previous displayed line
 				// - teststr.size() -> to account for the additional brackets put by DisplayOneLinerString()
 				// - 4 -> to account for adding 3 dots and a space to describe left out output.
+				if( term_width - current_disp_len - ( int )brackets.size() - ( int )continuation.size() < 0 )
+					continue;
+
 				opline = opline.substr( 0, term_width - current_disp_len - brackets.size() - continuation.size() );
 				opline += continuation;
 			}
+
+			if( !opline.empty() && * opline.begin() == ' ' )
+				opline.erase( opline.begin() );
 
 			MoveOutputCursorBack( prevdisp );
 
@@ -64,28 +75,30 @@ int DispExecute( std::string cmd, std::string & err, bool show_output )
 
 	MoveOutputCursorBack( prevdisp );
 
-	err = "";
+	if( create_temp_file ) {
+		err = "";
 
-	bool iserrfile = false;
-	std::fstream errfile;
-	errfile.open( TMP_FILE, std::ios::in );
-	if( errfile ) {
-		iserrfile = true;
-		std::string line;
-		while( std::getline( errfile, line ) )
-			err += line;
+		bool iserrfile = false;
+		std::fstream errfile;
+		errfile.open( TMP_FILE, std::ios::in );
+		if( errfile ) {
+			iserrfile = true;
+			std::string line;
+			while( std::getline( errfile, line ) )
+				err += line;
+		}
+		errfile.close();
+
+		if( iserrfile )
+			std::system( ( "rm -rf " + TMP_FILE ).c_str() );
 	}
-	errfile.close();
-
-	if( iserrfile )
-		std::system( ( "rm -rf " + TMP_FILE ).c_str() );
 
 	return pclose( pipe );
 }
 
-int DispExecuteNoErr( std::string cmd, bool show_output )
+int DispExecuteNoErr( std::string cmd, bool show_output, bool create_temp_file )
 {
 	std::string temperr;
 
-	return DispExecute( cmd, temperr, show_output );
+	return DispExecute( cmd, temperr, show_output, create_temp_file );
 }
